food_detection_bot/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
pydantic-settings==2.10.1
pytest==8.4.1
httpx==0.28.1
pytesseract==0.3.13
paddleocr==2.8.1
